Moving hierarchy functions from utils.py to hierarchy_utils.py.
