+++ nikola/nikola.py
@@ -61,7 +61,7 @@ from blinker import signal
 
 from .post import Post  # NOQA
 from .state import Persistor
-from . import DEBUG, filters, utils, shortcodes
+from . import DEBUG, filters, utils, hierarchy_utils, shortcodes
 from .plugin_categories import (
     Command,
     LateTask,
@@ -2085,7 +2085,7 @@ class Nikola(object):
         """Parse a category name into a hierarchy."""
         if self.config['CATEGORY_ALLOW_HIERARCHIES']:
             try:
-                return utils.parse_escaped_hierarchical_category_name(category_name)
+                return hierarchy_utils.parse_escaped_hierarchical_category_name(category_name)
             except Exception as e:
                 utils.LOGGER.error(str(e))
                 sys.exit(1)
@@ -2095,7 +2095,7 @@ class Nikola(object):
     def category_path_to_category_name(self, category_path):
         """Translate a category path to a category name."""
         if self.config['CATEGORY_ALLOW_HIERARCHIES']:
-            return utils.join_hierarchical_category_path(category_path)
+            return hierarchy_utils.join_hierarchical_category_path(category_path)
         else:
             return ''.join(category_path)
 
@@ -2120,7 +2120,7 @@ class Nikola(object):
             """Create category hierarchy."""
             result = []
             for name, children in cat_hierarchy.items():
-                node = utils.TreeNode(name, parent)
+                node = hierarchy_utils.TreeNode(name, parent)
                 node.children = create_hierarchy(children, node)
                 node.category_path = [pn.name for pn in node.get_path()]
                 node.category_name = self.category_path_to_category_name(node.category_path)
@@ -2131,7 +2131,7 @@ class Nikola(object):
 
         root_list = create_hierarchy(self.category_hierarchy)
         # Next, flatten the hierarchy
-        self.category_hierarchy = utils.flatten_tree_structure(root_list)
+        self.category_hierarchy = hierarchy_utils.flatten_tree_structure(root_list)
 
     @staticmethod
     def sort_posts_chronologically(posts, lang=None):
